# attendance_module_frontend/swagger3.json
{"openapi":"3.1.0","info":{"title":"Attendance Module","description":"For marking Attendance for Students","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"UserController","description":"User APIs"},{"name":"Role Controller","description":"Role Based APIs"},{"name":"Auth Controller","description":"Authentication APIs"},{"name":"StudentController","description":"Student APIs"},{"name":"CourseController","description":"Course APIs"}],"paths":{"/user/updateUser/{userId}":{"put":{"tags":["UserController"],"operationId":"updateUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatedUserRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}},"/student/update/{studentId}":{"put":{"tags":["StudentController"],"operationId":"updateStudent","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStudentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StudentResponseDto"}}}}}}},"/course/getcourse/{courseId}":{"put":{"tags":["CourseController"],"operationId":"updateCourse","parameters":[{"name":"courseId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CourseResponseDto"}}}}}}},"/attendance/{attendanceId}":{"put":{"tags":["attendance-controller"],"operationId":"update","parameters":[{"name":"attendanceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"query","required":true,"schema":{"type":"string","enum":["Present","Absent","Abscont"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Attendance"}}}}}},"delete":{"tags":["attendance-controller"],"operationId":"delete","parameters":[{"name":"attendanceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/attendance/by-student-name/{studentName}":{"put":{"tags":["attendance-controller"],"operationId":"updateByStudentName","parameters":[{"name":"studentName","in":"path","required":true,"schema":{"type":"string"}},{"name":"date","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"status","in":"query","required":true,"schema":{"type":"string","enum":["Present","Absent","Abscont"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Attendance"}}}}}}},"/attendance/by-student-id/{studentId}":{"put":{"tags":["attendance-controller"],"operationId":"updateByStudentId","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"date","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"status","in":"query","required":true,"schema":{"type":"string","enum":["Present","Absent","Abscont"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Attendance"}}}}}}},"/student/addStudent":{"post":{"tags":["StudentController"],"operationId":"createStudent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StudentResponseDto"}}}}}}},"/course/addcourse":{"post":{"tags":["CourseController"],"operationId":"createCourse","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CourseResponseDto"}}}}}}},"/attendance/by-name":{"post":{"tags":["attendance-controller"],"operationId":"markByName","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AttendanceResponse"}}}}}}},"/api/role":{"post":{"tags":["Role Controller"],"operationId":"createRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoleResponseDto"}}}}}}},"/api/auth/register":{"post":{"tags":["Auth Controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{}}}}}}},"/api/auth/login":{"post":{"tags":["Auth Controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{}}}}}}},"/user/viewUser":{"get":{"tags":["UserController"],"operationId":"viewCourse","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}}},"/user/viewUser/{roleId}":{"get":{"tags":["UserController"],"operationId":"viewByRoleId","parameters":[{"name":"roleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}}},"/user/userbycourName/{courseName}":{"get":{"tags":["UserController"],"operationId":"viewStudentByCourse","parameters":[{"name":"courseName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}}},"/student/trainer/{trainerUserId}/students":{"get":{"tags":["StudentController"],"operationId":"viewStudentsByTrainer","parameters":[{"name":"trainerUserId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentResponseDto"}}}}}}}},"/student/getStudent":{"get":{"tags":["StudentController"],"operationId":"viewStudents","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentResponseDto"}}}}}}}},"/student/getStudent/{studentId}":{"get":{"tags":["StudentController"],"operationId":"viewByStudentId","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentResponseDto"}}}}}}}},"/student/getStudent/gender/{studentGender}":{"get":{"tags":["StudentController"],"operationId":"viewStudentByGender","parameters":[{"name":"studentGender","in":"path","required":true,"schema":{"type":"string","enum":["Male","Female"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentResponseDto"}}}}}}}},"/student/getStudent/course/{course}":{"get":{"tags":["StudentController"],"operationId":"viewStudentByCourse_1","parameters":[{"name":"course","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentResponseDto"}}}}}}}},"/course/viewcourse":{"get":{"tags":["CourseController"],"operationId":"viewAllCourses","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseResponseDto"}}}}}}}},"/attendance":{"get":{"tags":["attendance-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AttendanceResponse"}}}}}}}},"/attendance/trainer-dashboard":{"get":{"tags":["attendance-controller"],"operationId":"getTrainerDashboard","parameters":[{"name":"date","in":"query","required":false,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardResponse"}}}}}}},"/attendance/by-date":{"get":{"tags":["attendance-controller"],"operationId":"getByDate","parameters":[{"name":"date","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AttendanceResponse"}}}}}}}},"/attendance/admin-dashboard":{"get":{"tags":["attendance-controller"],"operationId":"getAdminDashboard","parameters":[{"name":"date","in":"query","required":false,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardResponse"}}}}}}},"/api/viewAllRole":{"get":{"tags":["Role Controller"],"operationId":"viewAllRole","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoleResponseDto"}}}}}}}},"/user/deleteUser/{userId}":{"delete":{"tags":["UserController"],"operationId":"deleteUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/student/delete/{studentId}":{"delete":{"tags":["StudentController"],"operationId":"deleteStudent","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UpdatedUserRequestDto":{"type":"object","properties":{"userName":{"type":"string"},"email":{"type":"string"},"userDes":{"type":"string"},"roleId":{"type":"integer","format":"int64"}}},"UserResponseDto":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"userName":{"type":"string"},"email":{"type":"string"},"userDes":{"type":"string"},"roleName":{"type":"string"},"courseName":{"type":"string"}}},"UpdateStudentRequest":{"type":"object","properties":{"studentName":{"type":"string"},"email":{"type":"string"},"studentGender":{"type":"string","enum":["Male","Female"]},"studentDob":{"type":"string","format":"date"},"studentQualification":{"type":"string"},"address":{"type":"string"}}},"StudentResponseDto":{"type":"object","properties":{"studentId":{"type":"integer","format":"int64"},"studentName":{"type":"string"},"email":{"type":"string"},"studentGender":{"type":"string","enum":["Male","Female"]},"studentDob":{"type":"string","format":"date"},"courseName":{"type":"string"},"courseDuration":{"type":"integer","format":"int64"},"studentQualification":{"type":"string"},"address":{"type":"string"}}},"CourseRequestDto":{"type":"object","properties":{"courseName":{"type":"string"},"courseDuration":{"type":"integer","format":"int64"}}},"CourseResponseDto":{"type":"object","properties":{"courseId":{"type":"integer","format":"int64"},"courseName":{"type":"string"},"courseDuration":{"type":"integer","format":"int64"}}},"Attendance":{"type":"object","properties":{"attendanceId":{"type":"integer","format":"int64"},"studentId":{"type":"integer","format":"int64"},"trainerUserId":{"type":"integer","format":"int64"},"courseId":{"type":"integer","format":"int64"},"attendanceDate":{"type":"string","format":"date"},"attendanceStatus":{"type":"string","enum":["Present","Absent","Abscont"]},"trainerName":{"type":"string"}}},"StudentRequestDto":{"type":"object","properties":{"studentName":{"type":"string"},"email":{"type":"string"},"studentGender":{"type":"string","enum":["Male","Female"]},"studentDob":{"type":"string","format":"date"},"studentQualification":{"type":"string"},"address":{"type":"string"},"courseId":{"type":"integer","format":"int64"}}},"AttendanceRequest":{"type":"object","properties":{"studentId":{"type":"integer","format":"int64"},"studentName":{"type":"string"},"attendanceDate":{"type":"string","format":"date"},"attendanceStatus":{"type":"string","enum":["Present","Absent","Abscont"]}}},"AttendanceResponse":{"type":"object","properties":{"attendanceId":{"type":"integer","format":"int64"},"studentName":{"type":"string"},"trainerName":{"type":"string"},"courseName":{"type":"string"},"attendanceDate":{"type":"string","format":"date"},"attendanceStatus":{"type":"string","enum":["Present","Absent","Abscont"]}}},"RoleRequestDto":{"type":"object","properties":{"roleId":{"type":"integer","format":"int64"},"roleName":{"type":"string"}}},"RoleResponseDto":{"type":"object","properties":{"roleName":{"type":"string"}}},"UserRequestDto":{"type":"object","properties":{"userName":{"type":"string"},"email":{"type":"string"},"userDes":{"type":"string"},"userPassword":{"type":"string"},"roleId":{"type":"integer","format":"int64"},"courseId":{"type":"integer","format":"int64"}}},"AuthRequestDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"DashboardResponse":{"type":"object","properties":{"trainerId":{"type":"integer","format":"int64"},"trainerName":{"type":"string"},"totalStudents":{"type":"integer","format":"int64"},"totalPresent":{"type":"integer","format":"int64"},"totalAbsent":{"type":"integer","format":"int64"},"totalOngoing":{"type":"integer","format":"int64"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}